Added parse_webhook_payload() entry point:
Two APIs. The first accepts a webhook payload string and returns the Perl
structure, just a glorified JSON parser interface.

The second accepts a webhook payload string and a callback, and for each
message in there calls the callback with three parameters: event type, the
message, and the timestamp.

The second API is recommended because it will, in the future, support
streaming parsing for very big payloads.

Signed-off-by: Pedro Melo <melo@simplicidade.org>